requirements.txt
python-opencv
numpy
pillow
pycryptodome
shutil